use separate module for referencing leaflet & related plugins
